Update MixtureCraft
This is Mixture Craft my Extended Crafting System, The point is to build through getTile placements of ritualistic or machine like setups activated by key items, elements like fire and water also used. This is a small template of that modding system for modders to start with it definately makes crafting more interesting. A more in depth look can be found in my vampire mod.